Compile for linux aarch64
This allows me to run the code on my raspberry pi.

I would like to gather as much data as possible about the AIs, so I want
to leave it running for a few days. I don't want to leave my desktop on
the whole time, so I'm going to run it on my raspberry pi instead.